various prod fixes to get this semiworking

+++ cron/workspacesTransition.js
@@ -25,7 +25,8 @@ async function checkDBConsistency() {
        we see in the database. */
 
     const ec2 = new AWS.EC2();
-    const running_hosts = (await ec2.describeInstances({
+    const running_host_set = new Set();
+    const reservations = (await ec2.describeInstances({
         Filters: [
             {
                 Name: 'tag-key',
@@ -33,8 +34,12 @@ async function checkDBConsistency() {
             },
         ],
         MaxResults: 500,
-    }).promise()).data.Reservations.Instances.map(instance => instance.InstanceId);
-    const running_host_set = new Set(running_hosts);
+    }).promise()).Reservations;
+    for (const reservation of reservations) {
+	for (const instance of Object.values(reservation.Instances)) {
+	    running_host_set.add(instance.InstanceId);
+	}
+    }
 
     const db_hosts = (await sqldb.queryAsync(sql.select_nonstopped_workspace_hosts, [])).rows.map(instance => instance.instance_id);
     const db_hosts_set = new Set(db_hosts);
@@ -46,16 +51,21 @@ async function checkDBConsistency() {
                 diff.add(val);
             }
         }
+	return diff;
     };
 
     /* Kill off any host that is running but not in the db */
     const not_in_db = set_difference(running_host_set, db_hosts_set);
-    await sqldb.queryAsync(sql.add_terminating_hosts, { instances: Array.from(not_in_db) });
-    await ec2.terminateInstances({ InstanceIds: Array.from(not_in_db) }).promise();
+    if (not_in_db.size > 0) {
+	await sqldb.queryAsync(sql.add_terminating_hosts, { instances: Array.from(not_in_db) });
+	await ec2.terminateInstances({ InstanceIds: Array.from(not_in_db) }).promise();
+    }
 
     /* Any host that is in the db but not running we will mark as "terminated" */
     const not_in_ec2 = set_difference(db_hosts_set, running_host_set);
-    await sqldb.queryAsync(sql.set_terminated_hosts, { instances: Array.from(not_in_ec2) });
+    if (not_in_ec2.size > 0) {
+	await sqldb.queryAsync(sql.set_terminated_hosts, { instances: Array.from(not_in_ec2) });
+    }
 }
 
 async function terminateHosts() {
@@ -65,7 +75,9 @@ async function terminateHosts() {
         config.workspaceHostLaunchTimeoutSec,
     ];
     const hosts = (await sqldb.callAsync('workspace_hosts_find_terminable', params)).rows[0].terminated_hosts || [];
-    await ec2.terminateInstances({ InstanceIds: hosts }).promise();
+    if (hosts.length > 0) {
+	await ec2.terminateInstances({ InstanceIds: hosts }).promise();
+    }
 }
 
 async function checkHealth() {
@@ -79,10 +91,10 @@ async function checkHealth() {
         } catch (err) {
             healthy = false;
         }
-        if (response.statusCode !== 200) {
+        if (response && response.statusCode !== 200) {
             healthy = false;
         }
-        logger.info(`Host ${host.hostname} ($host.instance_id) is unhealthy!`);
+        logger.info(`Host ${host.hostname} (${host.instance_id}) is unhealthy!`);
 
         if (!healthy) {
             await sqldb.queryOneRowAsync(sql.set_host_unhealthy, { instance_id: host.instance_id });
